feat: introduce IXLFontEngine to make font library replaceable
Extract font operations from IXLGraphicEngine into a new IXLFontEngine
interface, allowing the underlying font library (SixLabors.Fonts) to be
swapped without replacing image handling.

- Add IXLFontEngine interface with the 5 font methods
- Add DefaultFontEngine wrapping existing SixLabors.Fonts 1.x logic
- Refactor DefaultGraphicEngine to delegate font work to DefaultFontEngine
- Add FontEngine property to LoadOptions and XLWorkbook
- Migrate internal consumers to use FontEngine instead of GraphicEngine
- DefaultGraphicEngine implements IXLFontEngine for backward compatibility
