+++ XLibur/Excel/LoadOptions.cs
@@ -15,6 +15,11 @@ public class LoadOptions
     public static IXLGraphicEngine? DefaultGraphicEngine { internal get; set; }
 
     /// <summary>
+    /// A font engine that will be used for workbooks without explicitly set font engine.
+    /// </summary>
+    public static IXLFontEngine? DefaultFontEngine { internal get; set; }
+
+    /// <summary>
     /// Should all formulas in a workbook be recalculated during a load? Default value is <c>false</c>.
     /// </summary>
     public bool RecalculateAllFormulas { get; set; } = false;
@@ -23,6 +28,13 @@ public class LoadOptions
     /// Graphic engine used by the workbook.
     /// </summary>
     public IXLGraphicEngine? GraphicEngine { get; set; }
+
+    /// <summary>
+    /// Font engine used by the workbook for text measurement and font metrics.
+    /// When not set, falls back to the font engine of the graphic engine (if it implements
+    /// <see cref="IXLFontEngine"/>), or the built-in <see cref="DefaultFontEngine"/>.
+    /// </summary>
+    public IXLFontEngine? FontEngine { get; set; }
 
     /// <summary>
     /// DPI for the workbook. Default is 96.
